app/requirements.txt
bcrypt==3.1.6
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pkg-resources==0.0.0
psycopg2-binary==2.7.7
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
Werkzeug==0.14.1
